+++ grid2op/tests/test_EnvironmentCpy.py
@@ -10,9 +10,7 @@
 # This module will test that the environment, when copied, works as expected (ie with making some basic tests
 # for the results of "env.copy()"
 
-import copy
-import pdb
-import time
+import unittest
 import warnings
 
 from grid2op.tests.helper_path_test import *