+++ grid2op/tests/test_MultiProcess.py
@@ -7,6 +7,7 @@
 # This file is part of Grid2Op, Grid2Op a testbed platform to model sequential decision making in power systems.
 
 import warnings
+import unittest
 from grid2op.tests.helper_path_test import *
 
 import grid2op