+++ grid2op/tests/test_Observation.py
@@ -11,6 +11,7 @@ import tempfile
 import warnings
 import copy
 import pdb
+import unittest
 
 from grid2op.tests.helper_path_test import *