+++ grid2op/tests/test_RulesByArea.py
@@ -7,6 +7,8 @@
 # This file is part of Grid2Op, Grid2Op a testbed platform to model sequential decision making in power systems.
 
 from itertools import chain
+import unittest
+
 from grid2op.tests.helper_path_test import *
 from grid2op.Exceptions import *
 from grid2op.Parameters import Parameters