+++ grid2op/tests/test_RunnerFast.py
@@ -26,6 +26,7 @@ warnings.simplefilter("error")
 
 class TestRunner(HelperTests, unittest.TestCase):
     def setUp(self):
+        super().setUp()
         self.init_grid_path = os.path.join(PATH_DATA_TEST_PP, "test_case14.json")
         self.path_chron = PATH_ADN_CHRONICS_FOLDER
         self.parameters_path = None