+++ grid2op/tests/test_alert_gym_compat.py
@@ -6,6 +6,8 @@
 # SPDX-License-Identifier: MPL-2.0
 # This file is part of Grid2Op, Grid2Op a testbed platform to model sequential decision making in power systems.
 
+import unittest
+
 from grid2op.tests.helper_path_test import *
 
 import grid2op